Splitting Tiles into Separate Files
In order to facilitate re-use of our Arc tiles and Task tiles currently only present in the Arc View Screen, I have removed the `arcTile` and `taskTile` definitions from `arc_view_screen.dart` and put them in `arc_tile.dart` and `task_tile.dart` respectively.
