+++ client/src/arcplanner/lib/src/ui/arc_tile.dart
@@ -1,0 +1,69 @@
+import 'package:flutter/material.dart';
+import '../blocs/bloc.dart';
+import 'drawer_menu.dart';
+import '../model/arc.dart';
+import '../model/task.dart';
+import 'package:auto_size_text/auto_size_text.dart';
+
+Widget arcTile(Arc arc, BuildContext context) {
+  var description = arc.description;
+  if (description == null) {
+    description = '';
+  }
+
+  return Container(
+    decoration: BoxDecoration(
+      border: Border(
+        bottom: BorderSide(
+          color: Colors.grey[350],
+        ),
+        top: BorderSide(
+          color: Colors.grey[350],
+        ),
+      ),
+    ),
+    height: MediaQuery.of(context).size.height * 0.20,
+    child: ListTile(
+      title: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        mainAxisAlignment: MainAxisAlignment.start,
+        children: <Widget>[
+          Container(
+            padding: EdgeInsets.only(
+              top: 10.0,
+              bottom: 10.0,
+            ),
+            child: AutoSizeText(arc.title,
+              style: TextStyle(
+                color: Colors.black,
+                fontWeight: FontWeight.bold,
+              ),
+              maxFontSize: 24.0,
+              minFontSize: 18.0,
+              maxLines: 1,
+              overflow: TextOverflow.ellipsis,
+            ),
+          ),
+          Container(
+            padding: EdgeInsets.only(
+              bottom: 10.0,
+            ),
+            child: AutoSizeText(description,
+              style: TextStyle(
+                color: Colors.black,
+              ),
+              maxFontSize: 16.0,
+              minFontSize: 12.0,
+              maxLines: 4,
+              overflow: TextOverflow.ellipsis,
+            ),
+          ),
+        ],
+      ),
+      onTap: () {
+        bloc.arcViewInsert({ 'object' : arc, 'flag': 'getChildren'});
+      } 
+      //onLongPress: ,
+    ),
+  );
+}
